# Micro optimizations