--- a/src/Microsoft.AspNet.Server.Kestrel/Http/Connection.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/Http/Connection.cs
@@ -238,7 +238,7 @@ namespace Microsoft.AspNet.Server.Kestrel.Http
             }
         }
 
-        private enum ConnectionState
+        private enum ConnectionState : byte
         {
             Open,
             Shutdown,