--- a/src/Microsoft.AspNet.Server.Kestrel/Http/Frame.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/Http/Frame.cs
@@ -900,11 +900,11 @@ namespace Microsoft.AspNet.Server.Kestrel.Http
             Log.ApplicationError(ex);
         }
 
-        private enum HttpVersionType
+        private enum HttpVersionType : byte
         {
-            Unknown = -1,
-            Http1_0 = 0,
-            Http1_1 = 1
+            Unknown,
+            Http1_0,
+            Http1_1
         }
     }
 }