--- a/tools/Microsoft.StandardsPolice/StandardsPoliceCompileModule.cs
+++ b/tools/Microsoft.StandardsPolice/StandardsPoliceCompileModule.cs
@@ -277,7 +277,7 @@ namespace Microsoft.StandardsPolice
         {
         }
 
-        private enum ClassZone
+        private enum ClassZone : byte
         {
             Ignored,
             BeforeStart,