I was playing around with Kestrel a little bit and decided to take a look at the code. I couldn't build in VS 2015 (Not using Update 1 RC), but could build via the command line. Please give me feedback on anything I could change or update. It's a learning experience playing with libuv :)

* Enums use byte instead of int. This will need to be benchmarked, ints may preform better on some cpus.
* Fixed an issue where a new buffer was being created when it didn't need to be.
* Don't write to streams or do extra work (allocations) if we don't need to be.