@niemyjski 2015-11-24
I see that there are other cases where were allocating ArraySegments. I'd need to dive into this more to see if we could save from allocating one or more of them. I do see cases where we might be able to save some additional time by checking to see if the buffer contains any data before calling write.