@halter73 2015-12-01
Thanks for the PR.

Generally speaking, we don't like to take perf changes that don't come with measurements showing either a request-per-second improvement or significant per-request heap allocation reduction.

I haven't measured the perf impact of the myself, but from looking at the changes I don't suspect this does either. Feel free to reopen if you have measurements that prove me wrong.