@niemyjski 2015-12-01
What do you recommend me doing to add tests for measurements? It's kinda hard to say this increases by x on my machine cause it can be different per machine..