@khellang 2015-12-01
@niemyjski You can establish a baseline on your machine, then apply the changes and measure the difference?