@niemyjski 2015-12-01
Do you have a test that outputs benchmarks I can then paste a before and after?

Sent from Outlook




On Tue, Dec 1, 2015 at 1:33 PM -0800, "Kristian Hellang" <notifications@github.com> wrote:












@niemyjski You can establish a baseline on your machine, then apply the changes and measure the difference?



—
Reply to this email directly or view it on GitHub.