@khellang 2015-12-01
Since these are "micro optimizations", I don't know how much measurable difference there will be running https://github.com/aspnet/benchmarks, but you could just use a simple stopwatch and do the same thing over and over for x iterations?