@halter73 2015-12-01
https://github.com/aspnet/benchmarks also shows how to use `wrk` to get measurements. If you have a linux machine/vm to run wrk on, that's probably a bit easier than setting up your own load test harness. Unfortunately, I also don't expect the changes I've seen made so far in this PR will make a measurable impact.