@niemyjski 2015-12-02
I ran these across different vms... Doing 32 threads on OS X caused some errors to be thrown by `wrk`.

before 
```
Running 10s test @ http://10.211.55.6:5001/plaintext
  26 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.12ms   11.84ms 115.67ms   79.08%
    Req/Sec     6.47k     1.66k   21.03k    78.77%
  1676891 requests in 10.07s, 211.10MB read
  Socket errors: connect 7, read 0, write 0, timeout 0
Requests/sec: 166514.06
Transfer/sec:     20.96MB
```
after
```
Please write a guide on how to use your own kestrel source... You can't expect people to test this if you make users jump through 10 hoops just to do a pull request. I tried updating references/adding it as an existing project reference, updating global.json/project.json without any luck after a half hour. I could get it to compile but runtime / assembly load errors all over the place...
```