@halter73 2015-12-02
Sorry for the frustration this has caused you.

First of all, it might be better to lower the connection count if you are seeing connect errors since those can mess with the results. As long as you use the same parameters for each run, you should be fine.

[Updating the global.json](https://github.com/aspnet/Home/wiki/Engineering-guidelines#local-debugging-of-cross-repo-dependencies) is the way I normally go about testing changes across repositories. I'm not sure why you ran into assembly load errors. Have you tried rerunning `dnu restore` in the root of both repos?

If that works for you, I'll be sure to add the `dnu restore` step to the engineering guidelines I linked to. If that doesn't work for you, feel free to add more details about the errors and we can try to figure it out.

The reason I closed this is because I don't expect you will see any measurable improvements from this PR, but I'm glad to try to help anyone who wants to get started contributing to Kestrel.