@niemyjski 2015-12-02
Is there any chance you can try running my changes as well really quick? I have other things running on my macbook (elastic search restores and other things that could be skewing the results)

To get it to work I had to do this
remove code from `spcm` file:

   public class StandardsPoliceCompileModule : Microsoft.StandardsPolice.StandardsPoliceCompileModule
    {
    }

Add `"../KestrelHttpServer/src"` to global.json and add the existing kestrel project to the solution