@halter73 2015-12-02
Doh! I know the problem you are running into, and this one is specific Kestrel.

You need to add `../KestrelHttpServer/tools` to your global.json too (though removing StandardsPoliceCompileModule also works). Maybe this should be added to the Kestrel README.

Though we usually like the submitter of perf changes to provide their own measurements first, I'll try testing the PR in our perf lab in the next batch of tests since it's been such a hassle for you.