@niemyjski 2015-12-09
Were you able to run these tests?