@AspNetSmurfLab 2015-12-09
__Baseline (origin/dev dff3a4f):__

	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    11.57ms   28.80ms 363.98ms   90.57%
	    Req/Sec    25.10k     3.54k   76.12k    94.73%
	  11993716 requests in 15.10s, 1.47GB read
	Requests/sec: 794311.34
	Transfer/sec:     99.99MB

	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    13.26ms   36.69ms 495.50ms   91.72%
	    Req/Sec    25.10k     3.23k   63.14k    90.65%
	  11973370 requests in 15.10s, 1.47GB read
	Requests/sec: 793022.60
	Transfer/sec:     99.83MB


	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     9.54ms   25.18ms 366.01ms   90.77%
	    Req/Sec    25.12k     2.92k   63.19k    91.38%
	  12014880 requests in 15.10s, 1.48GB read
	Requests/sec: 795755.29
	Transfer/sec:    100.17MB


__Merged (niemyjsk/dev):__

	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    14.21ms   33.14ms 379.55ms   87.93%
	    Req/Sec    24.92k     2.49k   48.33k    79.73%
	  11947374 requests in 15.10s, 1.47GB read
	Requests/sec: 791252.40
	Transfer/sec:     99.61MB


	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    13.27ms   38.14ms 602.15ms   89.61%
	    Req/Sec    25.07k     2.18k   41.86k    76.15%
	  12033096 requests in 15.10s, 1.48GB read
	Requests/sec: 796933.36
	Transfer/sec:    100.32MB


	Running 15s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    13.24ms   32.73ms 412.07ms   91.41%
	    Req/Sec    25.00k     2.32k   40.14k    77.99%
	  11993584 requests in 15.10s, 1.47GB read
	Requests/sec: 794309.38
	Transfer/sec:     99.99MB